Better token validation. Proper exit status on ENOENT, EACCES and SIGINT. Removed double prompt when sending SIGINT to a blocking process.

+++ source/src/include/defs.h
@@ -6,7 +6,7 @@
 /*   By: mtarrih <mtarrih@student.1337.ma>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/17 15:49:38 by mtarrih           #+#    #+#             */
-/*   Updated: 2025/09/21 23:15:59 by mtarrih          ###   ########.fr       */
+/*   Updated: 2025/09/26 20:22:25 by mtarrih          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ typedef struct s_redir {
 
 typedef struct s_cmd {
 	char **argv;
-	size_t argc;
+	int argc;
 	int stdin_fd;
 	int stdout_fd;
 	t_sllist *redirs;
